feat(consensus): CONS-401 — MessageBroadcaster + ValidatorMessage migrate to lib-consensus-core
Establishes `lib-consensus-core/src/ports/broadcaster.rs` as the single
canonical home for the `MessageBroadcaster` trait, satisfying the
CONS-401 acceptance criteria:
- `grep -rn "trait MessageBroadcaster" --include="*.rs"` returns
  exactly one match (the canonical definition).
- No re-exports of older trait shapes anywhere in the workspace.

The trait's signature takes `ValidatorMessage`, so the value type and
its transitive deps moved alongside.

## Moved to `lib-consensus-core/src/types/messages.rs`

- `ValidatorMessage` (3-variant enum)
- `ProposeMessage`, `VoteMessage`, `HeartbeatMessage`
- `ConsensusStateView`, `NetworkSummary` (heartbeat / vote payload)
- `Justification` (proposal previous-round vote bundle)

These were the last value types in `lib-consensus-net::validator_protocol`
that downstream crates needed; the move down the dep graph keeps the
port-trait + value-type pair colocated.

## Moved to `lib-consensus-core/src/ports/broadcaster.rs`

- `MessageBroadcaster` async trait with the full CE-ENG-1..7
  invariant docs preserved verbatim from the old location.

## Re-exports keep callers compiling

- `lib-consensus-net::validator_protocol`: re-exports the 7 message
  types from `lib_consensus_core::types::*` so existing import paths
  (`lib_consensus_net::validator_protocol::ValidatorMessage`, etc.)
  keep working unchanged.
- `lib-consensus::types`: replaces the in-place `MessageBroadcaster`
  trait definition with `pub use lib_consensus_core::ports::
  MessageBroadcaster`. Existing `lib_consensus::types::
  MessageBroadcaster` imports compile unchanged.

## Validation

- `cargo build --workspace` — clean.
- `cargo test -p lib-consensus-core --lib` — **41/41**.
- `cargo test -p lib-consensus --lib` — **134/134**.
- `cargo test -p lib-consensus-net --lib` — **53/53**.
- `cargo test -p lib-consensus-runtime --lib` — **15/15**.
- 41 + 134 + 53 + 15 = **243 tests pass**, no regressions.

## Acceptance against #2347

- ✅ `lib-consensus-core/src/ports/broadcaster.rs` is the single
  source of truth for `MessageBroadcaster`.
- ✅ `grep -rn "trait MessageBroadcaster"` returns exactly one match.
- ✅ No old trait shape re-exported under any path; both `lib-consensus`
  and `lib-consensus-net` use `pub use` redirects.

Closes #2347
